Carry the port as uint16_t end to end
A port was an int in the option structs and parsePort(), a short in
NetworkServer's constructor and NetworkClient::connect() (with a comment
explaining why ports above 32767 had to be cast back before the resolver
saw them), and an unsigned short out of NetworkServer::port(), so every
hand-off cast. It is a uint16_t everywhere now: the transport's
signatures, DEFAULT_PORT and parsePort(), ServerOptions, the client's
command line and ServerConnection, the test client's session and options,
and the test helpers; the casts and the comment are gone. The connect
menu's ImGui InputInt keeps its int scratch and converts once.

The render server, MPI server and remote viewer still keep their own
short and pass it to the same constructors, converting implicitly; their
declarations are outside this document's scope.

PR review fix-up 04 item 6.

+++ src/apps/interactive/scivisStudioRemote/client/Application.h
@@ -17,6 +17,7 @@
 // vsr_core
 #include "vsr/core/FlatMap.hpp"
 // std
+#include <cstdint>
 #include <functional>
 #include <memory>
 #include <string>
@@ -36,7 +37,7 @@ struct AddFileAnimationDatasetDialog;
 struct ClientCommandLine
 {
   std::string host{"127.0.0.1"};
-  int port{protocol::DEFAULT_PORT}; // 1..65535, as parsePort() accepts
+  uint16_t port{protocol::DEFAULT_PORT}; // 1..65535, as parsePort() accepts
   bool connectAtStartup{false};
 };
 
@@ -141,7 +142,7 @@ class Application : public vsr::ui::imgui::Application
   // Menu state //
 
   std::string m_host;
-  int m_port{0};
+  int m_port{0}; // the connect menu's InputInt edits it; 0..65535
   protocol::FrameEncoding m_preferredEncoding{protocol::FrameEncoding::Raw};
 
   // True while Lost: the panels show the frozen mirror but must not edit it.
